fix(analytics): cast EXTRACT(EPOCH) ROUND args to ::numeric in response-times (#2182)
Live prod E2E as Founder (Keychain smoke-login creds) hit HTTP 500 on
GET /api/analytics/response-times:
  function round(double precision, integer) does not exist

calculate_response_times() wrapped six duration metrics in
ROUND(AVG(EXTRACT(EPOCH ...)/86400), 2). EXTRACT(EPOCH ...) is double
precision and Postgres has no ROUND(double precision, int) — only
ROUND(numeric, int). Cast each metric to ::numeric before ROUND.

W89 class-audit of the module: calculate_sla_compliance and
calculate_revenue_metrics ROUND calls already cast their args to ::numeric
(the COUNT(*) FILTER(...)::numeric divisions) — response_times was the sole
offender. Same class as #2157 (crm resp-times) but a distinct endpoint the
earlier fix didn't cover.

Added static regression guard (test_historical_analytics_round_numeric.py):
paren-depth-parses every ROUND() body, asserts any EXTRACT(EPOCH ...) arg
routes through ::numeric. Proven to FAIL on the pre-fix source and PASS on
the fix (no live DB needed — runs in CI).

+++ apps/backend-rag/backend/services/analytics/historical_analytics.py
@@ -169,21 +169,23 @@ async def calculate_response_times(
                 COUNT(*) as total_practices,
 
                 -- Inquiry to Start
-                ROUND(AVG(EXTRACT(EPOCH FROM (p.start_date - p.inquiry_date)) / 86400), 2)
+                -- NOTE: EXTRACT(EPOCH ...) returns double precision; ROUND(double, int)
+                -- does not exist in Postgres, so cast the numeric arg to ::numeric.
+                ROUND((AVG(EXTRACT(EPOCH FROM (p.start_date - p.inquiry_date)) / 86400))::numeric, 2)
                     as avg_days_inquiry_to_start,
-                ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (p.start_date - p.inquiry_date)) / 86400), 2)
+                ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (p.start_date - p.inquiry_date)) / 86400))::numeric, 2)
                     as median_days_inquiry_to_start,
 
                 -- Start to Completion
-                ROUND(AVG(EXTRACT(EPOCH FROM (p.completion_date - p.start_date)) / 86400), 2)
+                ROUND((AVG(EXTRACT(EPOCH FROM (p.completion_date - p.start_date)) / 86400))::numeric, 2)
                     as avg_days_start_to_completion,
-                ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (p.completion_date - p.start_date)) / 86400), 2)
+                ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (p.completion_date - p.start_date)) / 86400))::numeric, 2)
                     as median_days_start_to_completion,
 
                 -- Total Cycle Time
-                ROUND(AVG(EXTRACT(EPOCH FROM (p.completion_date - p.inquiry_date)) / 86400), 2)
+                ROUND((AVG(EXTRACT(EPOCH FROM (p.completion_date - p.inquiry_date)) / 86400))::numeric, 2)
                     as avg_days_total_cycle,
-                ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (p.completion_date - p.inquiry_date)) / 86400), 2)
+                ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (p.completion_date - p.inquiry_date)) / 86400))::numeric, 2)
                     as median_days_total_cycle
 
             FROM practices p
